make update location independent from trace
In the acceptance ratio analysis,

q(x, y) = C(x, y) * weight (y `minus` x),

where C(x, y) is the probability of choosing the location of
the left-most difference between x and y. For the acceptance ratio
to be constant 1, I've assumed that C(x, y) = C(y, x), which is
only true if the location of update does not depend on the original
trace.