Modify Trace definition to accept initial mhPosteriorWeight as argument
This essentially modifies Trace to be a State monad w.r.t. mhPosteriorWeight,
where it previously was a Writer monad. It only changes implementation, while
the semantics of Trace remain the same.
This change is necessary since with CustomReal we can no longer use literal 1
for initial weight unless m is a MonadDist instance. This is not a problem
in practice, since we only ever use such m, but it makes it impossible
to define pure without assuming that m is MonadDist. Again this wouldn't be
a problem, but unfortunately it also means we can't make Trace a MonadTrans
instance, which is the real issue.